Added bar chart and line chart in dashboard and also get the API from individual employee
