+++ index.html
@@ -2,20 +2,15 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="apple-touch-icon" sizes="76x76" href="/src/style/assets/img/apple-icon.png" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <title>HWM</title>
-    <!-- <link href="src/style/index.css" rel="stylesheet" type="text/css"> -->
     <script>/* Pinegrow Interactions, do not remove */ (function(){try{if(!document.documentElement.hasAttribute('data-pg-ia-disabled')) { window.pgia_small_mq=typeof pgia_small_mq=='string'?pgia_small_mq:'(max-width:767px)';window.pgia_large_mq=typeof pgia_large_mq=='string'?pgia_large_mq:'(min-width:768px)';var style = document.createElement('style');var pgcss='html:not(.pg-ia-no-preview) [data-pg-ia-hide=""] {opacity:0;visibility:hidden;}html:not(.pg-ia-no-preview) [data-pg-ia-show=""] {opacity:1;visibility:visible;display:block;}';if(document.documentElement.hasAttribute('data-pg-id') && document.documentElement.hasAttribute('data-pg-mobile')) {pgia_small_mq='(min-width:0)';pgia_large_mq='(min-width:99999px)'} pgcss+='@media ' + pgia_small_mq + '{ html:not(.pg-ia-no-preview) [data-pg-ia-hide="mobile"] {opacity:0;visibility:hidden;}html:not(.pg-ia-no-preview) [data-pg-ia-show="mobile"] {opacity:1;visibility:visible;display:block;}}';pgcss+='@media ' + pgia_large_mq + '{html:not(.pg-ia-no-preview) [data-pg-ia-hide="desktop"] {opacity:0;visibility:hidden;}html:not(.pg-ia-no-preview) [data-pg-ia-show="desktop"] {opacity:1;visibility:visible;display:block;}}';style.innerHTML=pgcss;document.querySelector('head').appendChild(style);}}catch(e){console&&console.log(e);}})()</script>
   </head>
   <!-- <body class="text-gray-500"> -->
   <body class="m-0 font-sans antialiased font-normal text-base leading-default bg-gray-50 text-slate-500">  
     <div id="app"></div>
     <script type="module" src="/src/main.ts"></script>
-      <!-- plugin for charts  -->
-    <!-- <script src="/src/style/assets/js/plugins/chartjs.min.js" async></script> -->
-    <!-- plugin for scrollbar  -->
-    <!-- <script src="/src/style/assets/js/plugins/perfect-scrollbar.min.js" async></script> -->
-    <!-- main script file  -->
-    <!-- <script src="/src/style/assets/js/soft-ui-dashboard-tailwind.js?v=1.0.4" async></script> -->
   </body>
 </html>
