+++ src/style/assets/js/chart-1.js
@@ -8,7 +8,7 @@ new Chart(ctx, {
     labels: ["Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
     datasets: [
       {
-        label: "Sales",
+        label: "Avail",
         tension: 0.4,
         borderWidth: 0,
         borderRadius: 4,
